+++ src/module/data/pseudo-documents/advancements/_types.d.ts
@@ -9,6 +9,12 @@ declare module "./base-advancement.mjs" {
   }
 }
 
+declare module "./characteristic.mjs" {
+  export default interface CharacteristicAdvancement {
+
+  }
+}
+
 interface ItemGrantPool {
   uuid: string;
 }